+++ src/utils/str/ft_memcpy.c
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nle-gued <nle-gued@student.42lyon.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/18 16:06:46 by nle-gued          #+#    #+#             */
+/*   Updated: 2025/04/18 16:07:53 by nle-gued         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
+{
+	char		*d;
+	const char	*s = src;
+
+	d = dest;
+	while (n--)
+		*d++ = *s++;
+	return (dest);
+}
