+++ src/utils/str/ft_strdup_null.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strdup_null.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nle-gued <nle-gued@student.42lyon.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/18 16:08:14 by nle-gued          #+#    #+#             */
+/*   Updated: 2025/04/18 16:25:46 by nle-gued         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+/**
+ * @brief only difference with strdup: if s1 is NULL => returns ""
+ */
+char	*ft_strdup_null(char *s1)
+{
+	char	*dup;
+	size_t		len;
+
+	if (!s1)
+		len = 0;
+	else
+		len = ft_strlen(s1);
+	dup = malloc(len + 1);
+	if (!dup)
+		return (NULL);
+	ft_memcpy(dup, s1, len);
+	dup[len] = '\0';
+	return (dup);
+}
