+++ src/utils/str/ft_strjoin.c
@@ -1,70 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   str.c                                              :+:      :+:    :+:   */
+/*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mniemaz <mniemaz@student.42lyon.fr>        +#+  +:+       +#+        */
+/*   By: nle-gued <nle-gued@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/10 10:00:00 by mniemaz           #+#    #+#             */
-/*   Updated: 2025/04/17 10:29:56 by mniemaz          ###   ########.fr       */
+/*   Created: 2025/04/18 16:08:59 by nle-gued          #+#    #+#             */
+/*   Updated: 2025/04/18 16:23:48 by nle-gued         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	ft_putstr_fd(char *s, int fd)
-{
-	if (!s)
-		return ;
-	write(fd, s, ft_strlen(s));
-}
-
-int	ft_strlen(char *str)
-{
-	int	len;
-
-	len = 0;
-	while (str[len])
-		len++;
-	return (len);
-}
-
-void	*ft_memcpy(void *dest, const void *src, size_t n)
-{
-	char		*d;
-	const char	*s = src;
-
-	d = dest;
-	while (n--)
-		*d++ = *s++;
-	return (dest);
-}
-
-/**
- * @brief only difference with strdup: if s1 is NULL => returns ""
- */
-char	*ft_strdup_null(char *s1)
-{
-	char	*dup;
-	int		len;
-
-	if (!s1)
-		len = 0;
-	else
-		len = ft_strlen(s1);
-	dup = malloc(len + 1);
-	if (!dup)
-		return (NULL);
-	ft_memcpy(dup, s1, len);
-	dup[len] = '\0';
-	return (dup);
-}
-
 char	*ft_strjoin(char *s1, char *s2)
 {
 	char	*result;
-	int		len1;
-	int		len2;
+	size_t		len1;
+	size_t		len2;
 
 	if (!s1 && s2)
 		return (ft_strdup_null(s2));
