+++ src/utils/str/putstr_fd.c
@@ -1,12 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*   putstr_fd.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nle-gued <nle-gued@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/04/18 16:06:46 by nle-gued          #+#    #+#             */
-/*   Updated: 2025/04/18 16:06:47 by nle-gued         ###   ########.fr       */
+/*   Created: 2025/04/18 16:10:12 by nle-gued          #+#    #+#             */
+/*   Updated: 2025/04/18 16:11:09 by nle-gued         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "minishell.h"
+
+void	ft_putstr_fd(char *s, int fd)
+{
+	if (!s)
+		return ;
+	write(fd, s, ft_strlen(s));
+}
